requirements.txt
requests
feedparser
google-genai
markdown
beautifulsoup4
Pillow